+++ src/main/java/com/chrisrm/ideaddlc/icons/associations/AssociationsFactory.java
@@ -27,7 +27,6 @@
 package com.chrisrm.ideaddlc.icons.associations;
 
 import com.thoughtworks.xstream.XStream;
-import com.thoughtworks.xstream.io.xml.DomDriver;
 import org.jetbrains.annotations.NonNls;
 
 import java.net.URL;
@@ -43,26 +42,18 @@ public enum AssociationsFactory {
       "StaticMethodOnlyUsedInOneClass"})
   public static Associations create(final String associationsFile) {
     final URL associationsXml = AssociationsFactory.class.getResource(associationsFile);
-    @NonNls final XStream xStream = new XStream(new DomDriver());
+    @NonNls final XStream xStream = new XStream();
+    XStream.setupDefaultSecurity(xStream);
+    xStream.allowTypesByWildcard(new String[]{"com.chrisrm.ideaddlc.icons.associations.*"});
+
     xStream.alias("associations", Associations.class);
     xStream.alias("regex", RegexAssociation.class);
     xStream.alias("type", TypeAssociation.class);
-
-    final String psiClass = "com.intellij.psi.PsiClass";
-    if (isClass(psiClass)) {
-      xStream.alias("psi", PsiElementAssociation.class);
-    } else {
-      xStream.alias("psi", TypeAssociation.class);
-    }
 
     xStream.useAttributeFor(Association.class, "icon");
     xStream.useAttributeFor(Association.class, "name");
     xStream.useAttributeFor(RegexAssociation.class, "pattern");
     xStream.useAttributeFor(TypeAssociation.class, "type");
-
-    if (isClass(psiClass)) {
-      xStream.useAttributeFor(PsiElementAssociation.class, "type");
-    }
 
     try {
       return (Associations) xStream.fromXML(associationsXml);
@@ -71,18 +62,4 @@ public enum AssociationsFactory {
     }
   }
 
-  /**
-   * Tries to find whether a class is loaded by its className
-   *
-   * @param className the FQ classname
-   * @return true if found, otherwise false
-   */
-  public static boolean isClass(final String className) {
-    try {
-      Class.forName(className);
-      return true;
-    } catch (final ClassNotFoundException e) {
-      return false;
-    }
-  }
 }