+++ src/main/java/io/acari/DDLC/Constants.kt
@@ -8,8 +8,10 @@ const val DARK_MODE_CONSTANT: String = "Ḑ̶͍̲̳̫͗̚͠ä̵͉̠̬͙̤͚͆͆
 val UPDATE_MESSAGE: String = """
       What's New?<br>
       <ul>
-      <li>Fixed exception being thrown by Material UI integration, sorry! （ﾉ´д｀）</li>
+      <li>Doki-Doki Icons do not get overridden by other themes now.</li>
+      <li>More Kotlin icon associations.</li>
+      <li>Less icon residue when switching themes.</li>
       <br>
-      Thanks again for downloading <b>The Doki-Doki Theme</b>!<br>
+      Thanks again for downloading <b>The Doki-Doki Theme</b>! •‿•<br>
       <br>Please see the <a href="https://github.com/cyclic-reference/ddlc-jetbrains-theme/blob/master/docs/CHANGELOG.md">Changelog</a> for more details.
 """.trimIndent()